Improve the error handling in ack deadline extension (#2430)
* Improve the error handling in ack deadline extension

Wrap the `modify_ack_deadline` call with try exception. This call is to
extend the ack_deadline of the fetched messages while they are waiting
to be executed. This call is to prevent the message from being
redelivered.

In the case we get an error (e.g. 504) the modify_ack_deadline will
throw an exception and cause the thread to terminate. This fix is to
prevent the thread from terminating but carry on trying, so that if the
service is back to live, then the future fetched messages will be
extended with the new deadline. Else, it will cause redeliveries.

* Fix the formatting issue

* Remove whitespaces

---------

Co-authored-by: Asif Saif Uddin {"Auvi":"অভি"} <auvipy@gmail.com>
(cherry picked from commit a12f3a924a21a5ee1f3b5ee9f60dd93e11102c42)